+++ sources/com/google/zxing/common/detector/MathUtils.java
@@ -1,5 +1,6 @@
 package com.google.zxing.common.detector;
 
+/* loaded from: classes.dex */
 public final class MathUtils {
     private MathUtils() {
     }
@@ -11,13 +12,13 @@ public final class MathUtils {
     public static float distance(float aX, float aY, float bX, float bY) {
         float xDiff = aX - bX;
         float yDiff = aY - bY;
-        return (float) Math.sqrt((double) ((xDiff * xDiff) + (yDiff * yDiff)));
+        return (float) Math.sqrt((xDiff * xDiff) + (yDiff * yDiff));
     }
 
     public static float distance(int aX, int aY, int bX, int bY) {
         int xDiff = aX - bX;
         int yDiff = aY - bY;
-        return (float) Math.sqrt((double) ((xDiff * xDiff) + (yDiff * yDiff)));
+        return (float) Math.sqrt((xDiff * xDiff) + (yDiff * yDiff));
     }
 
     public static int sum(int[] array) {
