+++ sources/kotlin/jvm/internal/Ref.java
@@ -2,10 +2,12 @@ package kotlin.jvm.internal;
 
 import java.io.Serializable;
 
+/* loaded from: classes.dex */
 public class Ref {
     private Ref() {
     }
 
+    /* loaded from: classes.dex */
     public static final class ObjectRef<T> implements Serializable {
         public T element;
 
@@ -14,22 +16,25 @@ public class Ref {
         }
     }
 
+    /* loaded from: classes.dex */
     public static final class ByteRef implements Serializable {
         public byte element;
 
         public String toString() {
-            return String.valueOf(this.element);
+            return String.valueOf((int) this.element);
         }
     }
 
+    /* loaded from: classes.dex */
     public static final class ShortRef implements Serializable {
         public short element;
 
         public String toString() {
-            return String.valueOf(this.element);
+            return String.valueOf((int) this.element);
         }
     }
 
+    /* loaded from: classes.dex */
     public static final class IntRef implements Serializable {
         public int element;
 
@@ -38,6 +43,7 @@ public class Ref {
         }
     }
 
+    /* loaded from: classes.dex */
     public static final class LongRef implements Serializable {
         public long element;
 
@@ -46,6 +52,7 @@ public class Ref {
         }
     }
 
+    /* loaded from: classes.dex */
     public static final class FloatRef implements Serializable {
         public float element;
 
@@ -54,6 +61,7 @@ public class Ref {
         }
     }
 
+    /* loaded from: classes.dex */
     public static final class DoubleRef implements Serializable {
         public double element;
 
@@ -62,6 +70,7 @@ public class Ref {
         }
     }
 
+    /* loaded from: classes.dex */
     public static final class CharRef implements Serializable {
         public char element;
 
@@ -70,6 +79,7 @@ public class Ref {
         }
     }
 
+    /* loaded from: classes.dex */
     public static final class BooleanRef implements Serializable {
         public boolean element;
